+++ akshare/__init__.py
@@ -2395,9 +2395,10 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.9.77 fix: fix stock_hot_rank_wc interface
 1.9.78 fix: fix macro_usa_gdp_monthly interface
 1.9.79 fix: fix stock_sse_deal_daily interface
+1.9.80 fix: fix futures_spot_price interface
 """
 
-__version__ = "1.9.79"
+__version__ = "1.9.80"
 __author__ = "AKFamily"
 
 import sys