+++ docs/changelog.md
@@ -56,6 +56,10 @@
 
 ## 更新说明
 
+1.9.80 fix: fix futures_spot_price interface
+
+    1. 修复 futures_spot_price 接口
+
 1.9.79 fix: fix stock_sse_deal_daily interface
 
     1. 修复 stock_sse_deal_daily 接口
@@ -2294,6 +2298,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.9.80 fix: fix futures_spot_price interface
 
 1.9.79 fix: fix stock_sse_deal_daily interface